% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_usoGral_vR65_1.R
\name{ajp}
\alias{ajp}
\title{Funcion interna [ajp()]}
\usage{
ajp(fo, dat, svm.g, svm.c, rf.step, lg.family = "binomial", lg.step = TRUE,
  cart.meth = "class", rf.doBest = FALSE, rf.trace = FALSE,
  rf.plot = FALSE, salida = "completa", ...)
}
\description{
Ajuste Parametros para modelos de ML utilizados
}
\details{
Utilizada para ajustar simultanteamente parametros de funciones dentro de loop
 con los mismos datos de entrenamiento (SVM,RF,GLM)
}
\keyword{#@export}
\keyword{internal}

