% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM__vR71_2.R
\name{asm2}
\alias{asm2}
\title{ASM: Ajuste Simultaneo Modelos}
\usage{
asm2(frml, datFULL, columnaY, modelosML, parti = 10, propme = 0.75,
  predicts.out = FALSE, semillas = NULL, metadatos = NULL,
  error.resum = TRUE, moc = c("vm", "mp", "waauc"),
  cart.ctrl = rpart::rpart.control(cp = 0.01), lgt.stepaic = FALSE,
  costsvm = 1, kernel = "radial", svm.gamma = 0.5, ntree = 100,
  mtry = 2, b.mfinal = 100, uso.em = TRUE, em.alfa10 = 0.5,
  em.iter = 100, em.tol = 1e-05, cdb.dens = "density", ...)
}
\arguments{
\item{frml}{#gral}

\item{datFULL}{#gral}

\item{columnaY}{#gral}

\item{modelosML}{#gral}

\item{parti}{#gral}

\item{propme}{#gral}

\item{predicts.out}{#nuevo: controla salida DFs con Ypred x modelo}

\item{semillas}{#gral}

\item{metadatos}{#gral}

\item{error.resum}{#gral}

\item{cart.ctrl}{#cart}

\item{lgt.stepaic}{#glm-logit}

\item{costsvm}{#svm}

\item{kernel}{#svm}

\item{svm.gamma}{#svm}

\item{ntree}{#rf}

\item{mtry}{#rf}

\item{b.mfinal}{#boosting}

\item{uso.em}{#cdb-em}

\item{em.alfa10}{#cdb-em}

\item{em.iter}{#cdb-em}

\item{em.tol}{#cdb-em}

\item{cdb.dens}{#cdb-gral}

\item{...}{parametros adicionales para modelos ML}
}
\value{
objeto S3 de clase 'asm'
#@seealso gsub, grep
#@examples
asm2()
}
\description{
wrapper para ajuste simultaneo modelos ML
}
\keyword{asm}

