% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_aux_vR68_1.R
\name{auroc}
\alias{auroc}
\title{Titulo}
\usage{
auroc(obj, ...)
}
\arguments{
\item{obj}{Objeto}

\item{...}{No implementado}
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}

