% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_aux_vR68_1.R
\name{bestinfo}
\alias{bestinfo}
\title{Resumen de tiradas por bloque}
\usage{
bestinfo(lista, nloop, nombcols_dfl, qbest, indic, ...)
}
\arguments{
\item{lista}{Lista formada por tablas que en realidad son matrices de confusion}

\item{nloop}{Lista}

\item{nombcols_dfl}{Lista}

\item{qbest}{Lista}

\item{indic}{Lista}

\item{...}{No implementado}
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}

