% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM-CDB_dk_vR65_1.R
\name{dk}
\alias{dk}
\title{Densidad por nucleos para datos numericos o categoricos}
\usage{
dk(dfe, dft = NULL, vclase = NULL, dens = "density", ...)
}
\arguments{
\item{dfe}{Datos de entrenamiento}

\item{dft}{Datos de prueba/testing}

\item{vclase}{Variable categorica de clase}

\item{dens}{Funcion de densidad KDE implementada en R. Opciones: stats::density,
ks::kde, np::npudens}

\item{...}{No implementado}
}
\description{
Entra: datos (dfe,dft,variable clase),dens:funcion/paq usado para estimar f_KDE
Sale: f_KDE
}
\details{
(parrafos2-: van en Details, explayarse)
}

