% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM-CDB_EM_vR65_1.R
\name{emalg2k}
\alias{emalg2k}
\title{emalg2k: calculo alfas usando algoritmo EM y KDE}
\usage{
emalg2k(alfa1 = 0.5, datT, f1, f2, iter = 10, tol = 0.01, plota = FALSE,
  listalfa = FALSE)
}
\arguments{
\item{alfa1}{Valor inicial de \code{alfa1} para comenzar con algoritmo EM}

\item{datT}{Datos nuevos para hallar valores de alfa}

\item{f1, f2}{Funciones de densidad calculadas con densidades por nucleo (KDE)}

\item{iter}{Cantidad de iteraciones}

\item{tol}{Tolerancia}

\item{plota}{Grafico de valores obtenidos}

\item{listalfa}{Valor booleano, TRUE si}

\item{...}{parametros adicionales para modelos ML}
}
\value{
objeto S3 de clase 'list'
}
\description{
calcula valor de proporciones a priori/mezcla (alfa) mediante algoritmo EM
}
\examples{
emalg2k()
}
\keyword{emalg2k}

