% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_aux_vR68_1.R
\name{errp}
\alias{errp}
\title{Resumen de datos de matriz de confusion}
\usage{
errp(ldf)
}
\arguments{
\item{ldf}{Lista formada por tablas que en realidad son matrices de confusion}
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}

