% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_aux_vR68_1.R
\name{furoc}
\alias{furoc}
\title{Titulo}
\usage{
furoc(pred, patroro, med = NULL, ...)
}
\arguments{
\item{pred}{Pred}

\item{patroro}{Patroro}

\item{med}{Tipo de medida en eje vertical (OY), opciones "auc" (para area
debajo de curva ROC) o "tpr" (sensibilidad o tasa de verdaderos positivos)}

\item{...}{Argumentos adicionales pasados a funciones [ROCR::prediction()] o
[ROCR::performance()]}
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}

