% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM-CDB_aux_vR65_1.R
\name{gedn}
\alias{gedn}
\title{Titulo}
\usage{
gedn(n, prop, mu1, mu2, sigma1, sigma2, cl1 = "C1", cl2 = "C2",
  resu = FALSE)
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}
\keyword{internal}

