% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_usoGral_vR65_1.R
\name{infoxcol}
\alias{infoxcol}
\title{Funcion interna [infoxcol()]}
\usage{
infoxcol(dat, ...)
}
\description{
Extrae informacion util por columna de data.frame
}
\details{
Utilizada en generacion de Libro de Codiigos (LdC) de un data.frame cualquiera
}
\keyword{internal}

