% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_usoGral_vR65_1.R
\name{jyo_df}
\alias{jyo_df}
\title{Funcion interna [jyo_df()]}
\usage{
jyo_df(dfa, dfb, ..., ordensegun = 1, ordenado = TRUE)
}
\description{
Junta y ordena data.frame x ranking de primer df usado
}
\details{
Sirve para sacar informacion rankeable mas facilmente de objetos 'asm'.
Modificacion de https://www.r-statistics.com/2012/01/merging-two-data-frame-objects-while-preserving-the-rows-order/
}
\keyword{#@export}
\keyword{internal}

