% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_usoGral_vR65_1.R
\name{ldcmeta}
\alias{ldcmeta}
\title{Funcion interna [ldcmeta()]}
\usage{
ldcmeta(dat, nombrescol = c("NroVar", "NombreVar", "Class_R",
  "NivelesFactor"), exportar = TRUE)
}
\description{
Crea LdC "x arriba" con metadatos extraibles de R
}
\details{
Entra: dataframe, nombres de columnas (ej: ("NroVar","NombreVar","Class_R","NivelesFactor"));
Sale: version 'raw' de Libro de Codigos para datos en dataframe proporcionado
}
\keyword{internal}

