% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_aux_vR68_1.R
\name{mdcmed}
\alias{mdcmed}
\title{Genera medidas de resumen de matriz de confusion}
\usage{
mdcmed(tablamdc, dig = 3)
}
\arguments{
\item{tablamdc}{Lista formada por tablas que en realidad son matrices de confusion}

\item{dig}{Cantidad de digitos de redondeo}
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}

