% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_aux_vR68_1.R, R/fu_usoGral_vR65_1.R
\name{ord_df}
\alias{ord_df}
\title{Titulo}
\usage{
ord_df(df, compl = FALSE, rn = NULL, ...)

ord_df(df, compl = FALSE, rn = NULL, ...)
}
\arguments{
\item{compl}{compl}

\item{...}{No implementado}

\item{ldf}{Lista formada por tablas que en realidad son matrices de confusion}

\item{mdcmed}{mdcmed}

\item{modML}{modML}
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)

Funcion ord_df: ordena data.frame x ranking (>valor mas alto en el ranking)
}
\details{
(parrafos2-: van en Details, explayarse)

Sirve para sacar informacion rankeable mas facilmente de objetos 'asm'.
}
\keyword{#@export}
\keyword{internal}

