% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_metodos_vR71_2.R
\name{plot.asm2}
\alias{plot.asm2}
\title{ASM: Ajuste Simultaneo Modelos (super corta) -- metodos implementados
Metodo 'plot' para objetos de clase 'asm' (basico, completar)}
\usage{
\method{plot}{asm2}(x, tipo = "boxplot", destacar = NULL,
  col.dest = "blue", ind.mdc = c("sen", "fpr", "acc"), ...)
}
\arguments{
\item{x}{Objeto de clase 'asm'}

\item{tipo}{(x ahora implementado 'boxplot')}

\item{...}{(algo)}
}
\value{
#@seealso
#@examples
plot.asm()
}
\description{
ASM: Ajuste Simultaneo Modelos (super corta) -- metodos implementados
Metodo 'plot' para objetos de clase 'asm' (basico, completar)
}
\keyword{plot.asm}

