% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM-CDB_aux_vR65_1.R
\name{plot.cdb_dens}
\alias{plot.cdb_dens}
\title{Titulo -- DOCUMENTACION roxygen}
\usage{
\method{plot}{cdb_dens}(x, main = NULL, xlab = "x",
  ylab = "Funcion de densidad", type = "p", ...)
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)

#@export
}
\keyword{internal}

