% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM-CDB_aux_vR65_1.R
\name{predict.hist.x}
\alias{predict.hist.x}
\title{Titulo}
\usage{
\method{predict}{hist.x}(hh, x1)
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}
\keyword{internal}

