% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM-CDB_aux_vR65_1.R
\name{predict.nifM}
\alias{predict.nifM}
\title{Titulo}
\usage{
\method{predict}{nifM}(object, newdataM, ...)
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}
\keyword{internal}

