% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_usoGral_vR65_1.R
\name{reductable}
\alias{reductable}
\title{Funcion interna [reductable()]}
\usage{
reductable(tab)
}
\description{
Generalizacion de matrices de cofusion (MdConf) y calculos asociados
}
\details{
(En fase experimental; falta mejorar y generalizar para ampliar matrices
de cofusion)
}
\keyword{internal}

