% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_aux_vR68_1.R
\name{resblq}
\alias{resblq}
\title{Titulo}
\usage{
resblq(obj, formulas, binf.qbest = c(0.05, 0.1, 0.15, 0.2, 0.25, 0.5, 0.75,
  0.95), binf.indic = 10, ord.mdcmed, vdm.formula = NULL)
}
\arguments{
\item{obj}{Objeto}

\item{formulas}{Formulas}

\item{binf.qbest}{Cuantilas deseadas}

\item{binf.indic}{Sobre que indicador se requiere resumen (ver codigos en?)}

\item{ord.mdcmed}{Resultado de [ord_df()], es decir un ranking de?}

\item{vdm.formula}{Numero de formula (si objeto es una lista de formulas)}
}
\description{
(parrafo1: va debajo del titulo, que hace en 1 oracion)
}
\details{
(parrafos2-: van en Details, explayarse)
}

