% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_ASM_metodos_vR71_2.R
\name{summary.asm}
\alias{summary.asm}
\title{Metodo 'summary' para objetos de clase 'asm' (basico, completar)}
\usage{
\method{summary}{asm}(obj_1, obj_2 = NULL, nroform1, nroform2 = NULL,
  indic_MdC, mdr = "media")
}
\arguments{
\item{obj_1, obj_2}{Objetos de clase [asm()] para resumir/comparar}

\item{nroform1, nroform2}{Numero de formula utilizado (solo en caso de loops con muchas formulas)}

\item{indic_MdC}{Indicador(es) deseado(s) provenientes de una matriz de confusion. Las opciones
actuales son "todos","ROC","aciertos","errores"}

\item{mdr}{Medida de resumen a mostrar en salida. Opciones implementadas: "media", "mediana"}
}
\description{
(parrafo1: va debajo del titulo) Resume en tablas resultados de funcion \code{asm2()}
}
\details{
(parrafos2-: van en Details, explayarse)
}
\section{Comentarios adicionales (poner lo que quieras; esto no sale en help pero si lo ves en archivo R)}{

}

