% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fu_usoGral_vR65_1.R
\name{vespe}
\alias{vespe}
\title{Funcion interna [vespe()]}
\usage{
vespe(tdc, mostrar = TRUE, ...)
}
\description{
Calcula frecuencias esperadas en TDEs
}
\details{
Entra: tabla de contingencia/doble entrada; Sale: frecuencias esperadas
}
\keyword{internal}

